requirements.txt
av==16.1.0
keyboard==0.13.5
lz4==4.4.5
moderngl==5.12.0
numpy==2.4.2
pillow==12.1.0
protobuf==6.33.5
pyinstaller==6.19.0
PySide6==6.10.2
screeninfo==0.8.1
sounddevice==0.5.5
